Refactoring #3 Phase 1 - Part 3 [Server] Regrouping common contracts under the name of GenericModel
[Client] UI Update, improved the caching system by generalizing the functions. Prepared the terrain for a Excel and PDF Online support. Also fixed a bug during which you could have the Browser's previous user's rights - if it was an admin, you had the admin rights.

**** Note: The next changes will mark the beginning of the Phase 2. New features as well as other refactoring will be done.

+++ WorkstationClient/WorkstationBrowser/Controllers/LoginController.cs
@@ -68,7 +68,7 @@ namespace WorkstationBrowser.Controllers {
                 var authenticationManager = ctx.Authentication;
                 authenticationManager.SignIn(claimsIdentity);
 
-                _UserNotifications = newSession.WorkstationSession.GetAllNotifications(newSession.CurrentUser.id);
+                _UserNotifications = newSession.WorkstationSession.GetAllNotifications((int)newSession.CurrentUser.id);
                 _UserRights = RightsReader.Decode(newSession.CurrentUser.rights) as Dictionary<String,bool>;
 
                 Session.Add("HubInitialized", false);
